Create README.md
Just a quick script to get rid of values in a plist file with super-nested recurring instances of the value.

Intended use is to just run as a command line tool. So:

"formscrub myfile'

At present it just looks for a default value, because that's all I need it to do right meow.

Will soon add the capacity to rename the file on save, and to seek out and delete other potentially villainous data.